+++ game-hollowknight/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None
 
+## [2.0.1] - 2026-03-06
+
+- Fixed: Corrected some mod folder write checks for Xbox and GOG versions
+
 ## [2.0.0] - 2025-12-23
 
 - Initial release