+++ game-starsandisland/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None
 
+## [0.1.1] - 2026-03-06
+
+- Fixed: Changed to new BepInEx mod page
+- Changed: BepCfgMan is now downloaded from Nexus Mods
+
 ## [0.1.0] - 2026-02-18
 
 - Initial release